added strequ

+++ libft/libft/part 1/ft_strnstr.c
@@ -20,7 +20,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 	while (haystack[i] && len--)
 	{
 		j = 0;
-		while (needle[j] == haystack[i + j] && (j <= len))
+		while (needle[j] == haystack[i + j] && ((size_t)j <= len))
 		{
 			if (needle[j + 1] == 0)
 			{
